portal-console/backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlalchemy==2.0.43
psycopg[binary]==3.2.9
alembic==1.16.5
httpx==0.28.1
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
pydantic-settings==2.10.1
paramiko==3.5.1
python-multipart==0.0.20

